removed methods from app.rb to create separate modules
